Add test using thread
